refactor: move "URL" resource to separate project (#33)
