+++ src/VirusTotalCore.Urls/Models/UrlReportAttributes.cs
@@ -1,23 +1,22 @@
 using VirusTotalCore.Common.Models.Analysis;
 
-namespace VirusTotalCore.Models.Analysis.URL;
+namespace VirusTotalCore.Urls.Models;
 
 /// <summary>
 /// Information about URL.
 /// </summary>
-public class UrlReportAttributes
+public sealed class UrlReportAttributes
 {
     public required Dictionary<string, string> Categories { get; set; }
     public int FirstSubmissionDate { get; set; }
     /// <summary>
-    /// UTC timestamp representing last time the IP was scanned.
+    /// UTC timestamp representing last time the URL was scanned.
     /// </summary>
     public int LastAnalysisDate { get; set; }
     /// <summary>
-    /// Number of different results from this scans.
+    /// Number of different results from this scan.
     /// </summary>
     public required ScanAnalysisStats LastAnalysisStats { get; set; }
-    
     /// <summary>
     /// Result from URL scanners. dict with scanner name as key and a dict with notes/result from that scanner as value.
     /// Key presents engine name.
