+++ tests/VirusTotalCore.Tests/VirusTotalCore.Tests.csproj
@@ -31,6 +31,7 @@
       <ProjectReference Include="..\..\src\VirusTotalCore\VirusTotalCore.csproj" />
       <ProjectReference Include="..\..\src\VirusTotalCore.Common\VirusTotalCore.Common.csproj" />
       <ProjectReference Include="..\..\src\VirusTotalCore.IpAddresses\VirusTotalCore.IpAddresses.csproj" />
+      <ProjectReference Include="..\..\src\VirusTotalCore.Urls\VirusTotalCore.Urls.csproj" />
     </ItemGroup>
 
     <ItemGroup>
